+++ src/diffalign_viewer/__init__.py
@@ -1,3 +1,3 @@
 """diffalign result viewer."""
 
-__version__ = "0.1.3"
+__version__ = "0.1.4"